Fix error the pool gemfirePool already exists
Motivations:

java.lang.IllegalStateException: A pool named "gemfirePool" already exists
This is essentially a bean cyclic dependency caused by the pool-name
attribute of the client regions.

Changes:
The pool-name attribute of the client region "Transaction" is removed.
Others are not changed, as long as the problem seems to be gone.

Results:
Now the FastFootShoesRetailApp app starts up OK.